Add CLI for train.py (#1337)
* train cli

* precommit

* update train to reflect script

* rm extraneoud

* add eval back

* eval

* trim eval

* rm train

* fix repo

* fixing repo

* readme

* precommit

* test

* import typer

* typer

* use cmd line

* use cmd line

* fix llm-foundry train train to llm-foundry train

* test

* test

* test

* test

* test

* test

* test

* test

* test

* test

* test

* precommit

* deprecate old script to use new one

* readd main

* reformatting

* typo

* typo

* typo

* typo

* precommit

* precommit

* precommit

* precommit

* test

* test

* debug

* debug

* debug

* debug

* resolve 2/3 comments

* typo

* help

* rename

* import

* import pathing

* precommit

* precommit

* precommit

* rm type ignore

* rm List for list

* type ignore

* precommit

---------
