+++ llmfoundry/train/__init__.py
@@ -1,0 +1,17 @@
+# Copyright 2024 MosaicML LLM Foundry authors
+# SPDX-License-Identifier: Apache-2.0
+from llmfoundry.train.train import (
+    TRAIN_CONFIG_KEYS,
+    TrainConfig,
+    train,
+    train_from_yaml,
+    validate_config,
+)
+
+__all__ = [
+    'train',
+    'train_from_yaml',
+    'TrainConfig',
+    'TRAIN_CONFIG_KEYS',
+    'validate_config',
+]
